change(net): cleanups and renames
* move dial_peer onto MagicEndpoint
* some renames for clarity
* remove iroh_net::client module
