+++ iroh-net/src/lib.rs
@@ -1,9 +1,10 @@
 #![recursion_limit = "256"]
 
-pub mod client;
 pub mod defaults;
-pub mod endpoint;
+pub mod magic_endpoint;
 pub mod hp;
 pub mod net;
 pub mod tls;
 pub mod util;
+
+pub use magic_endpoint::MagicEndpoint;
